+++ R/plotDetectionBoxplot.R
@@ -38,9 +38,12 @@
 #'   as a single "year"
 #' @param effort if not \code{NULL}, a dataframe describing on effort times
 #'   to be formatted with \link{formatEffort}. If effort data is not provided
-#'   then times with zero detections will not be properly accounted for.
+#'   then times with zero detections will not be properly accounted for. 
+#'   Alternatively, if columns "effortStart" and "effortEnd" are present in
+#'   \code{x}, then these values will be used for start and end of effort
 #' @param dropZeroes logical flag to remove boxplots where all observations
 #'   are zero (these would normally appear as a flat line at zero)
+#' @param title optional title for the plot
 #' @param returnData if \code{TRUE} then no plot will be generated, instead the
 #'   dataframe that would normally be used to make the plot will be returned
 #'   
@@ -61,6 +64,7 @@ plotDetectionBoxplot <- function(x,
                                  combineYears=FALSE, 
                                  effort=NULL,
                                  dropZeroes=FALSE, 
+                                 title=NULL,
                                  returnData=FALSE) {
     binChoice <- c('hour', 'day', 'week', 'month')
     binSplit <- strsplit(bin, '/')[[1]]
@@ -71,7 +75,9 @@ plotDetectionBoxplot <- function(x,
     
     bigBin <- match.arg(binSplit[2], binChoice)
     smallBin <- match.arg(binSplit[1], binChoice)
-    
+    if(unitToPeriod(bigBin) <= unitToPeriod(smallBin)) {
+        stop('"bin" numerator must be smaller than denominator e.g. "day/week" not "week/day"')
+    }
     if(!is.null(facet) &&
        !facet %in% group) {
         # warning('"facet" must be included in "group"')
@@ -80,6 +86,10 @@ plotDetectionBoxplot <- function(x,
     missCol <- group[!group %in% names(x)]
     if(any(missCol)) {
         stop('Column(s) ', paste0(missCol, collapse=', '), ' are not in "x"')
+    }
+    if(is.null(effort) &&
+       all(c('effortStart', 'effortEnd') %in% names(x))) {
+        effort <- distinct(select(x, all_of(c('effortStart', 'effortEnd', group))))
     }
     for(col in group) {
         if(!col %in% names(effort)) {
@@ -115,7 +125,10 @@ plotDetectionBoxplot <- function(x,
         sumAcross <- c(sumAcross, 'year')
         switch(bigBin,
                'week' = {
-                   FUN <- isoweek
+                   FUN <- function(x) {
+                       isoweek(floor_date(x, unit='week'))
+                   }
+                   # FUN <- isoweek
                    levs <- 1:53
                    xLabs <- c('Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sept', 'Oct', 'Nov', 'Dec')
                    xAt <- 1 + c(0, 4, 8, 12, 17, 21, 25, 30, 34, 39, 43, 48)
@@ -250,7 +263,7 @@ plotDetectionBoxplot <- function(x,
         coord_cartesian(ylim=yLim)
     if(isTRUE(combineYears)) {
         g <- g +
-            scale_x_discrete(breaks=xAt, labels=xLabs, name=oneUp(bigBin))
+            scale_x_discrete(breaks=xAt, labels=xLabs, name=oneUp(bigBin), drop=FALSE)
     } else {
         g <- g +
             scale_x_datetime(date_labels = '%b-%Y', name='Date')
@@ -293,8 +306,13 @@ plotDetectionBoxplot <- function(x,
     groupTitle <- paste0(sumAcross, collapse=', ')
     groupTitle <- paste0('grouped across: ', groupTitle)
     binTitle <- paste0(oneUp(smallBin), 's per ', oneUp(bigBin))
-    title <- paste0(binTitle, ', ', groupTitle)
-    g <- g + ggtitle(title)
+    if(is.null(title)) {
+        title <- paste0(binTitle, ', ', groupTitle)
+        g <- g + ggtitle(title)
+    } else {
+        subtitle <- paste0(binTitle, ', ', groupTitle)
+        g <- g + ggtitle(title, subtitle = subtitle)
+    }
     g
 }